requirements.txt
Pillow
Requests
streamlit
streamlit_timeline
altair==4.2.2
